HLE: Fix integer sign inconcistency accross the codebase (#2222)
* Make all title id instances unsigned

* Replace address and size with ulong instead of signed types

Long overdue change.
Also change some logics here and there to optimize with the new memory
manager.

* Address Ac_K's comments

* Remove uneeded cast all around

* Fixes some others misalignment

+++ Ryujinx.Cpu/MemoryHelper.cs
@@ -8,28 +8,28 @@ namespace Ryujinx.Cpu
 {
     public static class MemoryHelper
     {
-        public static void FillWithZeros(IVirtualMemoryManager memory, long position, int size)
+        public static void FillWithZeros(IVirtualMemoryManager memory, ulong position, int size)
         {
             int size8 = size & ~(8 - 1);
 
             for (int offs = 0; offs < size8; offs += 8)
             {
-                memory.Write<long>((ulong)(position + offs), 0);
+                memory.Write<long>(position + (ulong)offs, 0);
             }
 
             for (int offs = size8; offs < (size - size8); offs++)
             {
-                memory.Write<byte>((ulong)(position + offs), 0);
+                memory.Write<byte>(position + (ulong)offs, 0);
             }
         }
 
-        public unsafe static T Read<T>(IVirtualMemoryManager memory, long position) where T : struct
+        public unsafe static T Read<T>(IVirtualMemoryManager memory, ulong position) where T : struct
         {
             long size = Marshal.SizeOf<T>();
 
             byte[] data = new byte[size];
 
-            memory.Read((ulong)position, data);
+            memory.Read(position, data);
 
             fixed (byte* ptr = data)
             {
@@ -37,7 +37,7 @@ namespace Ryujinx.Cpu
             }
         }
 
-        public unsafe static long Write<T>(IVirtualMemoryManager memory, long position, T value) where T : struct
+        public unsafe static ulong Write<T>(IVirtualMemoryManager memory, ulong position, T value) where T : struct
         {
             long size = Marshal.SizeOf<T>();
 
@@ -48,18 +48,18 @@ namespace Ryujinx.Cpu
                 Marshal.StructureToPtr<T>(value, (IntPtr)ptr, false);
             }
 
-            memory.Write((ulong)position, data);
+            memory.Write(position, data);
 
-            return size;
+            return (ulong)size;
         }
 
-        public static string ReadAsciiString(IVirtualMemoryManager memory, long position, long maxSize = -1)
+        public static string ReadAsciiString(IVirtualMemoryManager memory, ulong position, long maxSize = -1)
         {
             using (MemoryStream ms = new MemoryStream())
             {
                 for (long offs = 0; offs < maxSize || maxSize == -1; offs++)
                 {
-                    byte value = memory.Read<byte>((ulong)(position + offs));
+                    byte value = memory.Read<byte>(position + (ulong)offs);
 
                     if (value == 0)
                     {
